# Replace dynahash hash table in HNSW with simplehash, for speed