This speeds up HNSW index build by about 10% in some quick tests on my laptop. Should never be slower. Searches benefit too, especially with high ef_search settings.

For the sake of consistency, I also changed VACUUM to use it, but I don't think it's performance critical there.

(I did this a while ago already but got distracted and forgot about it before submitting a PR. I remembered it now)

## Benchmark

I used a subset of the glove-100-angular dataset from ann-benchmarks (I think that's what this was anyway; I extracted this test data a while ago and didn't keep notes):

```
postgres=# \d items_small
                Table "public.items_small"
  Column   |    Type     | Collation | Nullable | Default 
-----------+-------------+-----------+----------+---------
 id        | integer     |           |          | 
 embedding | vector(100) |           |          | 
Indexes:
    "items_small_embedding_idx" hnsw (embedding vector_cosine_ops) WITH (m='16', ef_construction='200')

postgres=# select count(*) from items_small;
 count 
-------
 50000
(1 row)
```

On master:
```
postgres=# \timing
Timing is on.
postgres=# reindex index items_small_embedding_idx ;
REINDEX
Time: 80313.248 ms (01:20.313)
postgres=# reindex index items_small_embedding_idx ;
REINDEX
Time: 72729.057 ms (01:12.729)
postgres=# reindex index items_small_embedding_idx ;
REINDEX
Time: 71265.495 ms (01:11.265)
postgres=# 
```

With this PR:
```
postgres=# \timing
Timing is on.
postgres=# reindex index items_small_embedding_idx ;
REINDEX
Time: 70778.870 ms (01:10.779)
postgres=# reindex index items_small_embedding_idx ;
REINDEX
Time: 64770.006 ms (01:04.770)
postgres=# reindex index items_small_embedding_idx ;
REINDEX
Time: 62349.657 ms (01:02.350)
```

Not a very rigorous test, but shows that this helps. I'm pretty confident this will never lose over dynahash, even if it might not make a difference with other workloads. The simplehash implementation and the murmurhash function I used here are simply better.