@hlinnaka 2023-12-15
There are some regression test failures, seem to be VACUUM related. I'm not seeing them locally, but I'll try to debug them later.