@ankane 2023-12-15
Thanks @hlinnaka! This looks promising for in-memory index build times from initial testing. I think we can improve performance for the pointer hash further in Postgres 13+ by storing the hash value on `HnswElement` and using `pointers_insert_hash` instead of `pointers_insert` to reduce hash calculations.