@ankane 2023-12-15
Pushed code for `pointers_insert_hash` to the [simplehash branch](https://github.com/pgvector/pgvector/compare/simplehash).

Also, re regression failures: looks like they were due to unintialized memory in `hash_tid` and are now fixed.