@hlinnaka 2023-12-18
> I think we can improve performance for the pointer hash further in Postgres 13+ by storing the hash value on HnswElement and using pointers_insert_hash instead of pointers_insert to reduce hash calculations.

Oh that's interesting! I can confirm that I see a significant gain from that in that same benchmark. I wouldn't have expected that, the hash function is pretty cheap.

I suspect it has to do with cache misses and prediction. I was playing with another patch earlier to add __builtin_prefetch() calls to HnswSearchLayer(), to prefetch the cache lines of the elements it's about to visit, and that showed similar performance gains that you got from adding 'hash' to HnswElement. I think referencing the hash field triggers speculative fetching of the cache line earlier, which avoids a cache stall later. Or something like that, I'm not that familiar with low-level CPU details..

Anyway, whatever the reason, that's a nice additional speedup! Thanks